requirements.txt
httpx>=0.27.0
xmltodict>=0.13.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
respx>=0.21.0
